open-ai-agent-sdk/rafay-agent/pyproject.toml
[project]
name = "Rafay-agent"
version = "0.1.0"
description = 'A simple agent that uses the OpenAI Agent SDK and GEMINI API KEY to respond to user input.'
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.95.1",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]

authors = [
    { name = "Abdul Rafay", email = "rafaynadeem77@gmail.com" }
]